websocket/requirements.test.txt
pytest
pytest-aiohttp
